Move app files to a folder and update import
